requirements.txt
sqlalchemy==2.0.31
python-dotenv==1.0.1
Flask==3.0.3
psycopg2==2.9.9